requirements.txt
fastapi
uvicorn
python-dotenv
transformers
torch
sentencepiece
elevenlabs
pydantic
jinja2
uuid